getting somewhere, but need to debug
